--- src/components/Textbook/WordList.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import WordDetail from './WordDetail';
-import Styles from './textbook.module.scss';
-
-const WordList = () => {
-  return <div>WordList</div>;
-};
-
-export default WordList;
