+++ src/components/Textbook/textbook.module.scss
@@ -2,26 +2,6 @@
 @import '../../styles/paragraph.scss';
 @import '../../styles/color.scss';
 
-@mixin btn {
-  display: flex;
-  align-items: center;
-  border-radius: 14px;
-  border: none;
-  padding: 0;
-  height: 44px;
-}
-@mixin btn-area {
-  display: flex;
-    div{
-      position: relative;
-      margin-right: -44px;
-      img{
-        width: 44px;
-        height: 44px;
-      }
-    }
-  
-}
 
 .textbook {
   display: flex;
@@ -30,142 +10,7 @@
   align-items: center;
   background-color: $greyLight;
 
-  // styles for heading
-
-  .heading{
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 1180px;
-    background-color: $white;
-    padding: 20px;
-    border-radius: 14px;;
-    .icon{
-      display: flex;
-      img{
-        padding-right: 12px;
-      }
-    }
-    .buttons{
-      display: flex;
-      gap:10px;
-      .sprint {
-        @include btn-area;
-        button{
-          @include btn;
-          padding-left: 53px;
-          padding-right: 45px;
-          letter-spacing: -0.64px;
-        }
-      }
-      .audio{
-        @include btn-area;
-        button{
-          @include btn;
-          padding-left: 59px;
-          padding-right: 37px;
-          letter-spacing: -0.64px;
-        }
-      }
-      .icon {
-        position: relative;
-        .setting{
-          display: flex;
-          width: 40px;
-          height: 40px;
-          align-items: center;
-          background-color: $greyLight;
-          border-radius: 4px;
-          svg {
-            margin: 0 auto; 
-            color: $greyIcon;
-          }
-          &:hover{
-            background-color: $cyanDark;
-            svg{
-              color:$white
-            }
-          } 
-        }
-        .dropdownContainer{
-          position: absolute;
-          width:max-content;
-          top:40px;
-          right: 0;
-          background-color: $white;
-          border-radius: 14px;
-          display: flex;
-          padding: 20px;
-          box-shadow: 0px 11.449064254760742px 11.449064254760742px 0px rgba(52, 41, 39, 0.04), 0px 11.449064254760742px 68.69439697265625px 0px rgba(52, 41, 39, 0.04), 0px 68.69439697265625px 68.69439697265625px 0px rgba(52, 41, 39, 0.04), 0px 91.59251403808594px 91.59251403808594px 0px rgba(52, 41, 39, 0.04), 0px 183.18502807617188px 183.18502807617188px 0px rgba(52, 41, 39, 0.12), 0px 343.4719543457031px 343.4719543457031px 0px rgba(52, 41, 39, 0.08);
-
-          ul{
-            display: flex;
-            flex-direction: column;
-            gap:20px;
-            li{
-              display: flex;
-              gap:10px;
-
-              .container {
-                position: relative;
-                padding-left: 30px;
-                cursor: pointer;
-              }
-              /* Hide the browser's default checkbox */
-              .container input {
-                position: absolute;
-                opacity: 0;
-                cursor: pointer;
-              }
-              
-              /* Create a custom checkbox */
-              .checkmark {
-                position: absolute;
-                top: 0;
-                left: 0;
-                height: 20px;
-                width: 20px;
-                background-color: $checkBox;
-                border: solid 1px $greyIcon;
-                border-radius: 4px;
-              }
-              /* On mouse-over, add a greyLight background color */
-              .container:hover input ~ .checkmark {
-                background-color: $greyLight;
-              }
-              
-              .container input:checked ~ .checkmark {
-                background-color: $cyanDark;
-              }
-              
-              .container .checkmark:after {
-                content: "";
-                position: absolute;
-                display: none;
-              }
-              
-              .container input:checked ~ .checkmark:after {
-                display: block;
-              }
-              
-              .container .checkmark:after {
-                left: 6px;
-                top: 1px;
-                width: 6px;
-                height: 12px;
-                border: solid white;
-                border-width: 0 2px 2px 0;
-                transform: rotate(45deg);
-              }  
-            }
-          }
-        }
-      }
-    }
-  }
-
-  // styles for subHeading
-
+  // styles for SubTitle
   .subheading{
     margin-top: 40px;
     margin-bottom: 40px;
@@ -194,124 +39,6 @@
     }
   }
   
-  // styles for word detail
-  .wordDetail{
-    display: flex;
-    border-radius: 14px;
-    background-color: $white;
-    margin-bottom: 40px;
-    width: 926px;
-    box-shadow: 0px 22.898128509521484px 45.79625701904297px -11.449064254760742px rgba(52, 41, 39, 0.08);
-    .leftImage {
-      position: relative;
-      .image{
-        border-top-left-radius: 14px;
-        border-bottom-left-radius: 14px;
-        height: 100%;
-        width:337px;
-        max-width: none;
-        object-fit: cover;
-      }
-      .buttonContainer {
-        display: flex;
-        flex-direction: column;
-        gap: 8px;
-        position: absolute;
-        border-top-left-radius: 8px;
-        border-bottom-left-radius: 8px;
-        top: 40px;
-        right: 0;
-        bottom:auto;
-        justify-content: flex-end;
-        padding: 8px;
-        background-color: #000000B2;
-        transition: opacity 0.3s ease;
-        color: $white;
-
-      button{
-        display: flex;
-        gap:8px;
-        img{
-          width: 16px;
-          height: 16px;
-          margin-top: 3px;
-        }
-        .hidden{
-          display: none;
-        }
-      }
-      &:hover .hidden {
-        display: block;
-      }
-    }
-  }
-    
-    .rightDetail{
-      padding: 40px 33px 40px 40px;
-      display: flex;
-      flex-direction: column;
-      gap:20px;
-      .word{
-        display: flex;
-        flex-direction: column;
-        gap:12px;
-
-        .wordName{
-          display: flex;
-          justify-content: space-between;
-          h3{
-            display: flex;
-            gap:12px;
-          }
-          .synonyms{
-            color:$greyDark;
-          }
-          .play{
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            width: 40px;
-            height: 40px;
-            background-color: $greyLight;
-            border-radius: 4px;
-            svg {
-              color: $greyIcon;
-              width: 24px;
-              height: 24px;
-            }
-            &:hover {
-              background-color: $cyanDark;
-              svg {
-                color: $white;
-              }
-            }
-          }
-        }
-        p {
-          letter-spacing: 0.04em;
-          color: $black;
-        }
-      }
-      .explain{
-        display: flex;
-        flex-direction: column;
-        gap:12px;
-        p{
-          font-weight: 500;
-          line-height: 150%;
-        }
-        .definition{
-          color:$black;
-        }
-      }
-      .buttons{
-        display: flex;
-        gap:20px;
-      }
-      
-    }
-  }
-
   // style for empty
   .empty{
     display: flex;
@@ -335,7 +62,7 @@
     }
   }
 
-  // style for pageNumber
+  // style for PageNumber
   .pageNumber{
     display: flex;
     gap:24px;
